+++ Api/src/Player/Service/ActionModifierServiceInterface.php
@@ -8,5 +8,5 @@ interface ActionModifierServiceInterface
 {
     public function getAdditiveModifier(Player $player, array $scopes, array $types, ?string $target = null): int;
 
-    public function getMultiplicativeModifier(Player $player, array $scopes, array $types, ?string $target = null): int;
+    public function getMultiplicativeModifier(Player $player, array $scopes, array $types, ?string $target = null): float;
 }
